+++ riverdriver/riversqlite/migration/main/003_river_job_tags_non_null.up.sql
@@ -2,3 +2,5 @@
 -- Normally `river_job.tags` is set to `NOT NULL` with a `DEFAULT` here, but since
 -- SQLite was added well after 003 came about, we push that to version 006 index.
 --
+
+SELECT 1;